Значение было недопустимо малым или недопустимо большим для Int32. #321

+++ VkNet/Model/Career.cs
@@ -42,7 +42,7 @@ namespace VkNet.Model
 		/// <summary>
 		/// Год окончания работы.
 		/// </summary>
-		public int? Until { get; set; }
+		public ulong? Until { get; set; }
 
 		/// <summary>
 		/// Должность.
